fix(ui): prevent circular import
