+++ ui/src/utils/tags.ts
@@ -36,12 +36,12 @@ const regexpString = [
   LINK_TAG_REGEX.source, // single-word labeled url
 ];
 
-const TAG_REGEX = new RegExp(`${regexpString.join("|")}`, "mi");
+export const TAG_REGEX = new RegExp(`${regexpString.join("|")}`, "mi");
 
-const RAW_URL_REGEX =
+export const RAW_URL_REGEX =
   /\b(?:(?:[a-z][\w-]+:(?:\/{1,3}|[a-z0-9%])|www\d{0,3}[.]|[a-z0-9.-]+[.][a-z]{2,4}\/)(?:[^\s()<>]+|\(([^\s()<>]+|(\([^\s()<>]+\)))*\))+(?:\(([^\s()<>]+|(\([^\s()<>]+\)))*\)|[^\s`!()[\]{};:'".,<>?«»“”‘’]))/;
 
-const TAG_OR_RAW_URL_REGEX = new RegExp(
+export const TAG_OR_RAW_URL_REGEX = new RegExp(
   `(${[...regexpString, RAW_URL_REGEX.source].join("|")})`,
   "mi"
 );
@@ -56,82 +56,3 @@ export const isMentionTag = (text: string): boolean =>
 export const isLinkTag = (text: string): boolean => LINK_TAG_REGEX.test(text);
 
 export const isRawUrl = (text: string): boolean => RAW_URL_REGEX.test(text);
-
-export const splitMewTextIntoContentParts = (
-  text: string,
-  links: LinkTarget[]
-): MewContentPart[] => {
-  const parts = text
-    .split(TAG_OR_RAW_URL_REGEX)
-    .filter((part) => part !== undefined && part.length > 0);
-
-  const contentParts = parts.map((part) => {
-    const partIsTag = isTag(part);
-
-    if (isRawUrl(part)) {
-      return {
-        text: part,
-        href: part,
-        tagType: MewTagType.RawUrl,
-      } as MewContentPart;
-    } else if (
-      links.length > 0 &&
-      partIsTag &&
-      part[0] === TAG_SYMBOLS.MENTION
-    ) {
-      const link = links.pop();
-
-      const agentPubKey = (link as MentionLinkTarget)[
-        LinkTargetName.Mention
-      ] as AgentPubKey;
-
-      const route: RouteLocationRaw = {
-        name: ROUTES.mention,
-        params: {
-          tag: part.substring(1),
-          agentPubKey: encodeHashToBase64(agentPubKey),
-        },
-      };
-
-      return {
-        text: part,
-        route,
-        tagType: MewTagType.Mention,
-      };
-    } else if (links.length > 0 && partIsTag && part[0] === TAG_SYMBOLS.LINK) {
-      const link = links.pop();
-
-      return {
-        text: part,
-        href: (link as UrlLinkTarget).Url,
-        tagType: MewTagType.Link,
-      };
-    } else if (partIsTag && part[0] === TAG_SYMBOLS.HASHTAG) {
-      const route: RouteLocationRaw = {
-        name: ROUTES.hashtag,
-        params: { tag: part.substring(1) },
-      };
-
-      return {
-        text: part,
-        route,
-        tagType: MewTagType.Hashtag,
-      };
-    } else if (partIsTag && part[0] === TAG_SYMBOLS.CASHTAG) {
-      const route: RouteLocationRaw = {
-        name: ROUTES.cashtag,
-        params: { tag: part.substring(1) },
-      };
-
-      return {
-        text: part,
-        route,
-        tagType: MewTagType.Cashtag,
-      };
-    } else {
-      return { text: part } as MewContentPart;
-    }
-  });
-
-  return contentParts;
-};
